+++ tests/trio108.py
@@ -544,8 +544,9 @@ async def foo_func_5():  # error: 0, "exit", Statement("yield", lineno+2)
             ...
 
 
-# Completely ignore checking asynccontextmanagers, since they checkpoint on both
-# __aenter__ and __aexit__
+# Completely ignore checking asynccontextmanagers, since they might checkpoint on
+# either or both of __aenter__ and __aexit__ and correct context managers only
+# actually yield once (even if there might be multiple yield statements).
 @asynccontextmanager
 async def foo_cm_1():
     while True:
@@ -554,6 +555,7 @@ async def foo_cm_1():
 
 @contextlib.asynccontextmanager
 async def foo_cm_2():
+    # Just imagine some fancy control flow so that only one of these executes.
     yield
     yield
     yield
@@ -565,6 +567,7 @@ async def foo_cm_3():
         yield
 
 
+# note: not asynccontextmanager, so treat this function as an async iterable.
 @anything.contextmanager
 async def foo_cm_4():  # error: 0, "exit", Stmt("yield", line+1)
     yield  # error: 4, "yield", Statement("function definition", lineno-1)